requirements.txt
flake8
flake8-json
scipy
scikit-learn
joblib
pytorch
matplotlib
gzip
numpy
networkx
tqdm